project separate login and register tests
